I finished my blog unfortuntally without async (vue) maybe in future

+++ app/Http/Controllers/Profile/DeleteController.php
@@ -4,16 +4,18 @@ namespace App\Http\Controllers\Profile;
 
 use App\Http\Controllers\Controller;
 
+use App\Models\User;
 use App\Models\UserProfile;
+use App\Policies\ProfilePolicy;
 use Illuminate\Http\RedirectResponse;
-use Illuminate\Support\Facades\Storage;
-use Illuminate\View\View;
 
-class DeleteController extends Controller
+class DeleteController extends BaseController
 {
     public function __invoke(UserProfile $profile): RedirectResponse
     {
+        $this->authorize('delete', $profile);
         $profile->delete();
+        User::destroy($profile->user_id);
         return redirect()->route('login');
     }
